requirements.txt
streamlit
streamlit-chat
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
langchain-openai
chromadb
sentence-transformers
python-dotenv
pypdf
python-docx
unstructured[html,md]
ollama